Struct sev_serial::SerialPort

pub struct SerialPort { /* private fields */ }

Basic implementation that allows for writing to a serial port using the SEV-ES and SEV-SNP GHCB IOIO protocol, or using direct port-based IO, depending on which IO port factory is used in the wrapper enum.

See section 4.1.2 in https://www.amd.com/system/files/TechDocs/56421-guest-hypervisor-communication-block-standardization.pdf for more information on the GHCB IOIO protocol.

Implementations

impl SerialPort

pub unsafe fn new( base_address: u16, io_port_factory: PortFactoryWrapper ) -> Self

Creates a new instance of a serial port with the given base address.

Safety

This function is unsafe as callers must make sure that the base address represents a valid serial port device.

pub fn init(&mut self) -> Result<(), &'static str>

Initializes the serial port for writing.

We don’t require interrupts or FIFO, and don’t configure a maximum speed.

pub fn wait_for_empty_output(&mut self) -> Result<(), &'static str>

Wait until the output buffer is empty.

pub fn send(&mut self, data: u8) -> Result<(), &'static str>

Sends a byte of data via the serial port.